emailservice/cron_service/requirements.txt
confluent-kafka==1.7.0
psycopg2-binary==2.9.3
schedule==1.1.0
pydantic==1.8.2
kafka-python==2.0.2